// models/enums.ts
export enum ElementType {
  Grass = 'Grass',
  Fire = 'Fire',
  Water = 'Water',
  Lightning = 'Lightning',
  Psychic = 'Psychic',
  Fighting = 'Fighting',
  Darkness = 'Darkness',
  Metal = 'Metal',
  Dragon = 'Dragon',
  Fairy = 'Fairy',
  Colorless = 'Colorless',
  Ice = 'Ice',
  Poison = 'Poison',
  Ground = 'Ground',
  Flying = 'Flying',
  Bug = 'Bug',
  Rock = 'Rock',
  Ghost = 'Ghost',
}

export enum Supertype {
  Pokemon = 'Pokémon',
  Trainer = 'Trainer',
  Energy = 'Energy'
}

export enum Subtype {
  Basic = 'Basic',
  Stage1 = 'Stage 1',
  Stage2 = 'Stage 2',
  VMAX = 'VMAX',
  Radiant = 'Radiant'
}

export enum TrainerType {
  Item = 'Item',
  Supporter = 'Supporter',
  Stadium = 'Stadium',
  Tool = 'Pokémon Tool'
}

export enum Rarity {
  Common = 'Common',
  Uncommon = 'Uncommon',
  Rare = 'Rare',
  DoubleRare = 'Double Rare',
  UltraRare = 'Ultra Rare',
  SecretRare = 'Secret Rare',
  IllustrationRare = 'Illustration Rare',
}

export enum HoloPattern {
  None = 'None',
  Starlight = 'Starlight',
  Cosmos = 'Cosmos',
  Tinsel = 'Tinsel',
  Sheen = 'Sheen',
  CrackedIce = 'Cracked Ice',
  Crosshatch = 'Crosshatch',
  WaterWeb = 'Water Web',
  Sequin = 'Sequin',
  Pixel = 'Pixel',
  VerticalBars = 'Vertical Bars',
  BorderGlow = 'Border Glow'
}